refactor(value): share value alignment helpers
