+++ nemo_rl/models/value/workers/megatron_value_worker.py
@@ -80,6 +80,10 @@ from nemo_rl.models.policy.workers.base_policy_worker import AbstractPolicyWorke
 from nemo_rl.models.policy.workers.patches import apply_transformer_engine_patch
 from nemo_rl.models.value.config import ValueConfig
 from nemo_rl.models.value.interfaces import ValueOutputSpec
+from nemo_rl.models.value.utils import (
+    gather_and_right_shift_values,
+    right_shift_values,
+)
 from nemo_rl.utils.nsys import wrap_with_nvtx_name
 
 TokenizerType = TypeVar("TokenizerType", bound=PreTrainedTokenizerBase)
@@ -199,14 +203,7 @@ def _value_loss_prepare_fn(
     # Drop the value head's trailing singleton ([..., 1]).
     if values.ndim > 2 and values.shape[-1] == 1:
         values = values.squeeze(-1)
-    cp_size = (
-        1
-        if context_parallel_group is None
-        else torch.distributed.get_world_size(context_parallel_group)
-    )
-    if cp_size > 1:
-        values = allgather_cp_sharded_tensor(values, context_parallel_group, seq_dim=1)
-    values = torch.cat([torch.zeros_like(values[:, :1]), values[:, :-1]], dim=1)
+    values = gather_and_right_shift_values(values, context_parallel_group)
     values = values[:, : data["returns"].shape[1]]
     return {"logits": values}, data
 
@@ -720,10 +717,7 @@ class MegatronValueWorkerImpl(AbstractPolicyWorker):
                     )
                 else:
                     # [B, S]: shift right by 1 so values[t] = V(state before token t).
-                    values = output_tensor.squeeze(-1)
-                    values = torch.cat(
-                        [torch.zeros_like(values[:, :1]), values[:, :-1]], dim=1
-                    )
+                    values = right_shift_values(output_tensor.squeeze(-1))
                 return torch.tensor(0.0, device=values.device), {"values": values}
 
             return output_tensor, collection_fn
