+++ pyrefly.toml
@@ -203,6 +203,7 @@ project-includes = [
   "nemo_rl/models/policy/workers/patches.py",
   "nemo_rl/models/value/__init__.py",
   "nemo_rl/models/value/config.py",
+  "nemo_rl/models/value/utils.py",
   "nemo_rl/models/value/workers/__init__.py",
   "nemo_rl/utils/__init__.py",
   "nemo_rl/utils/checkpoint_engines/__init__.py",
